THU 04/06

+++ B20 Ex03 Lior 316266055 Shahar 204351845/Ex03.GarageLogic/Truck.cs
@@ -155,7 +155,7 @@ Truck Information
 Carrying Hazardous Goods: {2}
 Load Capacity: {3}",
                 VehicleToString(),
-                eNumberOfWheels.Truck,
+                (int)eNumberOfWheels.Truck,
                 isCarryingHazardousGoods,
                 m_LoadCapacity);
